Update for Controller files
As I decide on parts and find more information I will be updating the relevant folders in the Controller directory. For now, I have removed the datasheets for parts I am not using, add data sheets for parts I am using, and added papers related to FCC rules for transmitters. As always, my thoughts are in notes.txt

DR