Adding design files for controller v5, and controller v6.
Controller v5 has holes milled in the center of the board (which costs
substantially more). Controller v6 uses the board outline to create these holes
(there are no holes in the middle, just a cut into the board). I've added all of
the design files that I submitted to 4PCB and passed through FREEDFM.

In addition, I've added the CAM file job that I used to create these files. It
generates a GERBER file and EPS file of the board layout. 4PCB wanted the GERBER
File, but I found that the EPS file was easier for me two view. They contain
identical information.