Motors Pin configuration change
There was another another change in the motor pin outputs. The problem
was that there was a motor pin output in PB7 when the bootloder also
defines this pin as an output and turns it on when a program is being
downloaded, and we definitely don't want to have a motor spinning each
time we download a new program.

Also The led that indicates when to quadcopter is armed or disarmed was
changed to pin PB6.